+++ flinkboot-core/src/main/java/module-info.java
@@ -15,8 +15,8 @@ module io.github.sekelenao.flinkboot.core {
     exports io.github.sekelenao.flinkboot.core.api.resource;
     exports io.github.sekelenao.flinkboot.core.api.typing.time;
     exports io.github.sekelenao.flinkboot.core.api.typing.collection;
-    // Internal annotation exported specifically to flinkboot-kafka for code generation metadata
-    exports io.github.sekelenao.flinkboot.core.internal.annotation to io.github.sekelenao.flinkboot.kafka;
+    // Internal annotation exported specifically to companion modules for code generation metadata
+    exports io.github.sekelenao.flinkboot.core.internal.annotation to io.github.sekelenao.flinkboot.kafka, io.github.sekelenao.flinkboot.fluss;
 
     requires com.fasterxml.jackson.annotation;
     requires com.fasterxml.jackson.databind;