+++ flinkboot-fluss/src/main/java/io/github/sekelenao/flinkboot/fluss/api/properties/source/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Configuration properties for Apache Fluss sources in Apache Flink.
+ */
+package io.github.sekelenao.flinkboot.fluss.api.properties.source;